1、修改loadMapWithLock方法使其能同時判斷多個快取是否存在 2、將likeComment返回值改為Integer

+++ secret-article/src/main/java/com/shijiawei/secretblog/article/controller/AmsCommentController.java
@@ -75,9 +75,9 @@ public class AmsCommentController {
      */
 
     @PostMapping("/{articleId}/comments/{commentId}/likes")
-    public R<Long> likeComment(@NotNull @PathVariable(value = "articleId")  Long articleId,@NotNull @PathVariable(value = "commentId") Long commentId){
+    public R<Integer> likeComment(@NotNull @PathVariable(value = "articleId")  Long articleId,@NotNull @PathVariable(value = "commentId") Long commentId){
         log.debug("commentId:{}",commentId);
-        Long newLikes = amsCommentService.likeComment(articleId,commentId);
+        Integer newLikes = amsCommentService.likeComment(articleId,commentId);
 
         return R.ok(newLikes);
     }
